+++ haas/plugins/discoverer.py
@@ -22,18 +22,6 @@ from .i_discoverer_plugin import IDiscovererPlugin
 logger = logging.getLogger(__name__)
 
 
-def _is_in_package(module_name):
-    package = module_name.split('.', 1)[0]
-    try:
-        __import__(package)
-    except ImportError as exc:
-        exc_str = str(exc)
-        # Nasty! But fixes #98.
-        if exc_str == 'No module named {0}'.format(package):
-            return False
-    return True
-
-
 def _is_import_error_test(test):
     return isinstance(test, ModuleImportError)
 
@@ -378,11 +366,6 @@ class Discoverer(IDiscovererPlugin):
         logger.debug('Loading tests from %r', module_name)
         try:
             module = get_module_by_name(module_name)
-        except ImportError:
-            if not _is_in_package(module_name):
-                # Non-package import in Python 2.7; we ignore and continue
-                return self._loader.create_suite()
-            test = _create_import_error_test(module_name)
         except Exception:
             test = _create_import_error_test(module_name)
         else:
@@ -396,6 +379,10 @@ class Discoverer(IDiscovererPlugin):
     def _discover_tests(self, start_directory, top_level_directory, pattern):
         for curdir, dirnames, filenames in os.walk(start_directory):
             logger.debug('Discovering tests in %r', curdir)
+            dirnames[:] = [
+                dirname for dirname in dirnames
+                if os.path.exists(os.path.join(curdir, dirname, '__init__.py'))
+            ]
             for filename in filenames:
                 filepath = os.path.join(curdir, filename)
                 if not match_path(filename, filepath, pattern):