+++ CHANGES.rst
@@ -2,8 +2,8 @@
  ``haas`` CHANGELOG
 ====================
 
-Changes since version 0.6.2
-===========================
+Version 0.7.0
+=============
 
 Behaviour changes
 -----------------